fix-back : refactoring ; sonarqube 정적 분석을 토대로 npe가 발생할 수 있는 부분 위주로 수정합니다.

See merge request s08-webmobile2-sub2/S08P12A807!337

+++ togetdog/src/main/java/com/ssafy/togetdog/user/model/dto/UserInfoRespDTO.java
@@ -46,7 +46,7 @@ public class UserInfoRespDTO {
 
 		double rating = 0;
 		if (user.getRatingCount() != 0) {
-			rating = user.getRatingSum() / user.getRatingCount();
+			rating = (int) user.getRatingSum() / user.getRatingCount();
 		}
 		return UserInfoRespDTO.builder()
 				.email(user.getEmail())
